Amending execution and exit functions
