+++ 05_Execution/single_command_utils2.c
@@ -13,99 +13,152 @@
 #include "../includes/minishell.h"
 
 /**
- * @function: clean_up_function
- * @brief: freeing up resources when child process exits
+ * @function: handle_redirections_file_opening
+ * @brief: handling opening of files when redirections are called
+ 	and handling errors if files are not valid
  * 
  * @param t_redirect_single_command_params *params : structure to
  	store parameters for handling redirects / no redirects
- 	char ***env : *** is called in calling function
- 		only needed ** for freeing in child process
- 	char *input : initial readline from main function
- 		used for freeing purposes if program fails
+ 	***env: *** is called in the calling function
+ 	needed ** to free data if child process exits.
  * 
  * @return: void function
  */
 
-void	clean_up_function(t_redirect_single_command_params *params,
-			char ***env, char *input)
-
+void	handle_redirections_file_opening(
+			t_redirect_single_command_params *params, char ***env)
 {
-	free(input);
-	free_pipes(params->pipes, params->pipe_count);
-	free_dup_envp(*env);
-	rl_clear_history();
-}
-
-/**
- * @function: exiting_conditions_nonzero_pipecount
- * @brief: handling conditions to exit the
- 	child process if certain conditions are met
- 	if there are no heredocs
- * 
- * @param t_redirect_single_command_params *params : structure to
- 	store parameters for handling redirects / no redirects
- 	char ***env : *** is called in calling function
- 		only needed ** for freeing in child process
- 	char *input : initial readline from main function
- 		used for freeing purposes if program fails
- * 
- * @return: void function
- */
-
-void	exiting_conditions_nonzero_pipecount(t_redirect_single_command_params
-*params, char ***env, char *input)
-
-{
-	while (params->result->redirect[params->loop_counter] != NULL)
-		params->loop_counter++;
-	if ((ft_strcmp(params->result->cmd[0], "echo") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "cd") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "pwd") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "export") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "unset") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "env") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "exit") == 0)
-		|| (ft_strcmp(params->result->redirect[params->loop_counter - 1],
-				"a") == 0))
+	while (params->k < params->i)
 	{
-		write(2, "exiting redirects\n", 18);
-		clean_up_function(params, env, input);
-		exit(EXIT_SUCCESS);
+		if (ft_strcmp(params->result->redirect[params->k], "a") == 0)
+		{
+			params->k++;
+			continue ;
+		}
+		else if (ft_strcmp(params->result->redirect[params->k], "<") == 0)
+			params->input_fd = open(params->result
+					->rd_arg[params->rd_arg_counter], O_RDONLY);
+		else if (ft_strcmp(params->result->redirect[params->k], ">") == 0)
+			params->output_fd = open(params->result->rd_arg
+				[params->rd_arg_counter], O_WRONLY | O_CREAT | O_TRUNC, 0644);
+		else if (ft_strcmp(params->result->redirect[params->k], ">>") == 0)
+			params->output_fd = open(params->result->rd_arg
+				[params->rd_arg_counter], O_WRONLY | O_CREAT | O_APPEND, 0644);
+		if (params->input_fd < 0 || params->output_fd < 0)
+		{
+			ft_printf("%s: ", params->result->rd_arg[params->rd_arg_counter]);
+			perror("Error opening file");
+			clean_up_function(params, env);
+			exit(EXIT_FAILURE);
+		}
+		params->rd_arg_counter++;
+		params->k++;
 	}
 }
 
 /**
- * @function: exiting_conditions_zero_pipecount
- * @brief: handling conditions to exit the
- 	child process if certain conditions are met
- 	if there are heredocs
+ * @function: handle_dup_and_closing_fd
+ * @brief: handling dup function calls for both input and output fd
  * 
  * @param t_redirect_single_command_params *params : structure to
  	store parameters for handling redirects / no redirects
- 	char ***env : *** is called in calling function
- 		only needed ** for freeing in child process
- 	char *input : initial readline from main function
- 		used for freeing purposes if program fails
+ 	***env: *** is called in the calling function
+ 	needed ** to free data if child process exits.
  * 
  * @return: void function
  */
 
-void	exiting_conditions_zero_pipecount(t_redirect_single_command_params
-*params, char ***env, char *input)
-
+void	handle_dup_and_closing_fd(t_redirect_single_command_params
+*params, char ***env)
 {
-	if ((ft_strcmp(params->result->cmd[0], "echo") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "cd") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "pwd") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "export") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "unset") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "env") == 0)
-		|| (ft_strcmp(params->result->cmd[0], "exit") == 0)
-		|| (ft_strcmp(params->result->redirect[params->loop_counter],
-				"a") == 0))
+	if (params->input_fd > 0)
 	{
-		write(2, "exiting redirects for no pipe\n", 30);
-		clean_up_function(params, env, input);
+		if (dup2(params->input_fd, STDIN_FILENO) == -1)
+		{
+			perror("Dup2 failed for child for input_fd");
+			clean_up_function(params, env);
+			exit(EXIT_FAILURE);
+		}
+		close(params->input_fd);
+	}
+	if (params->output_fd > 0)
+	{
+		if (dup2(params->output_fd, STDOUT_FILENO) == -1)
+		{
+			perror("Dup2 failed for child for input_fd");
+			clean_up_function(params, env);
+			exit(EXIT_FAILURE);
+		}
+		close(params->output_fd);
+	}
+}
+
+/**
+ * @function: executing_execve_redirections
+ * @brief: executing execve function if redirections are called
+ * 
+ * @param t_redirect_single_command_params *params : structure to
+ 	store parameters for handling redirects / no redirects
+ 	***env: *** is called in the calling function
+ 	needed ** to free data if child process exits.
+ * 
+ * @return: void function
+ */
+
+void	executing_execve_redirections(t_redirect_single_command_params
+*params, char ***env)
+{
+	write(2, "Entering the redirect loop\n", 27);
+	if (params->input_fd == 0)
+	{
+		write(2, "I should not come here\n", 23);
+		if (dup2(params->pipes[params->pipe_count - 1][0], STDIN_FILENO) == -1)
+		{
+			perror("dup2 failed");
+			clean_up_function(params, env);
+			exit(EXIT_FAILURE);
+		}
+	}
+	write(2, "I should come here\n", 19);
+	if (access(params->result->cmd[0], F_OK) == 0)
+		params->command_path = params->result->cmd[0];
+	else
+		params->command_path = find_command
+			(&params->result->cmd[0], 0, *env);
+	if (execve(params->command_path, params->result->cmd, *env) == -1)
+	{
+		if (params->command_path != params->result->cmd[0])
+			free(params->command_path);
+		clean_up_function(params, env);
+		exit(EXIT_FAILURE);
+	}
+}
+
+/**
+ * @function: handle_execve_for_redirections
+ * @brief: handling execve conditions when there is redirections
+ * 
+ * @param t_redirect_single_command_params *params : structure to
+ 	store parameters for handling redirects / no redirects
+ 	***env: *** is called in the calling function
+ 	needed ** to free data if child process exits.
+ * 
+ * @return: void function
+ */
+
+void	handle_execve_for_redirections(t_redirect_single_command_params *params,
+	char ***env)
+{
+	write(2, "Entering the redirect loop\n", 27);
+	if (params->result->cmd[0] == NULL)
+	{
+		write(2, "No commands found, so we just exit\n", 35);
+		clean_up_function(params, env);
 		exit(EXIT_SUCCESS);
 	}
+	if (params->pipe_count > 0)
+		exiting_conditions_nonzero_pipecount(params, env);
+	else
+		exiting_conditions_zero_pipecount(params, env);
+	executing_execve_redirections(params, env);
 }
