+++ 05_Execution/single_command_utils4.c
@@ -13,60 +13,134 @@
 #include "../includes/minishell.h"
 
 /**
- * @function: checking_if_pipes_exist
- * @brief: checks through the nodes if pipe commands exist
- 	return to calling function if pipe commands are found
- 	as current function only handles single commands
-	this function is created for single_command.c
+ * @function: executing_execve_other_cases
+ * @brief: checks to be conducted before execve
+ 	is executed to fulfill certain test scenarios
  * 
- * @param t_list *node: structure of the linked list
- 	where it points to void content
+ * @param t_redirect_single_command_params *params : structure for
+ 	single_command parameters
+ 	***env: *** is called in the calling function
+ 	needed ** to free data if child process exits.
  * 
- * @return: -1 if pipes are found in the linked list structure
- 	    0  if no pipes are found
+ * @return: void function
  */
 
-int	checking_if_pipes_exist(t_list *node)
-
+void	executing_execve_other_cases(t_redirect_single_command_params
+*params, char ***env)
 {
-	t_list		*traverse;
-	t_exec_node	*result;
-
-	traverse = node;
-	while (traverse)
+	if (params->result->redirect != NULL)
 	{
-		result = (t_exec_node *)traverse->content;
-		if (result->type == AST_PIPE)
+		if ((ft_strcmp(params->result->redirect[params->loop_counter],
+					"a") == 0)
+			|| (ft_strcmp(params->result->redirect[params->loop_counter],
+					"<") == 0)
+			|| (ft_strcmp(params->result->redirect[params->loop_counter],
+					">") == 0)
+			|| (ft_strcmp(params->result->redirect[params->loop_counter],
+					">>") == 0))
 		{
-			ft_printf("Pipes detected. Proceeding to function with pipes\n");
-			return (-1);
+			ft_printf("exiting no redirects for no_pipes\n");
+			clean_up_function(params, env);
+			exit(EXIT_SUCCESS);
 		}
-		traverse = traverse->next;
 	}
-	return (0);
+	if (access(params->result->cmd[0], F_OK) == 0)
+		params->command_path = params->result->cmd[0];
+	else
+		params->command_path = find_command(&params->result->cmd[0], 0, *env);
+	if (execve(params->command_path, params->result->cmd, *env) == -1)
+	{
+		perror("execve error");
+		if (params->command_path != params->result->cmd[0])
+			free(params->command_path);
+		clean_up_function(params, env);
+		exit(EXIT_FAILURE);
+	}
 }
 
 /**
- * @function: handle_redirects
- * @brief: handling redirections if command consists of redirects
- 	execute_child_process is in single_command_utils1.c
+ * @function: handle_exit_conditions_other_cases
+ * @brief: handling conditions to exit child process
  * 
- * @param t_redirect_single_command_params *params : structure to
- 	store parameters for handling redirects / no redirects
- 	char ***env : *** is called in calling function
- 		only needed ** for freeing in child process
- 	char *input : initial readline from main function
- 		used for freeing purposes if program fails
+ * @param t_redirect_single_command_params *params : structure for
+ 	single_command parameters
+ 	***env: *** is called in the calling function
+ 	needed ** to free data if child process exits.
  * 
- * @return: -1 if forking fails, 0 if all commands is executed
+ * @return: void function
  */
 
-int	handle_redirects(t_redirect_single_command_params *params,
-	char ***env, char *input)
-
+void	handling_exit_conditions_other_cases(t_redirect_single_command_params
+*params, char ***env)
 {
-	while (params->result->redirect[params->i] != NULL)
-		params->i++;
+	if (params->result->cmd[0] == NULL)
+	{
+		ft_printf("No commands found, so we just exit\n");
+		clean_up_function(params, env);
+		exit(EXIT_SUCCESS);
+	}
+	params->loop_counter = 0;
+	while (params->result->redirect[params->loop_counter] != NULL)
+		params->loop_counter++;
+	if ((ft_strcmp(params->result->redirect[params->loop_counter - 1],
+				"a") == 0)
+		|| (ft_strcmp(params->result->redirect[params->loop_counter - 1],
+				"<") == 0)
+		|| (ft_strcmp(params->result->redirect[params->loop_counter - 1],
+				">") == 0)
+		|| (ft_strcmp(params->result->redirect[params->loop_counter - 1],
+				">>") == 0))
+	{
+		ft_printf("exiting no redirects\n");
+		clean_up_function(params, env);
+		exit(EXIT_SUCCESS);
+	}
+}
+
+/**
+ * @function: child_process_other_cases
+ * @brief: after a fork is created, having checks for
+ 	heredocs and deciding what to execute
+ * 
+ * @param t_redirect_single_command_params : structure for
+ 	single_command parameters
+ 	***env: *** is called in the calling function
+ 	needed ** to free data if child process exits.
+ * 
+ * @return: void function
+ */
+
+void	child_process_other_cases(
+			t_redirect_single_command_params *params, char ***env)
+{
+	ft_printf("Entering the no redirect loop\n");
+	if (params->pipe_count == 0)
+	{
+		write(2, "Entering here\n", 14);
+		executing_execve_other_cases(params, env);
+	}
+	else
+	{
+		write(2, "Entering here1\n", 15);
+		handling_exit_conditions_other_cases(params, env);
+	}
+}
+
+/**
+ * @function: handle_fork_plus_executing_child
+ * @brief: forks the process and creates a child process to handle it
+ * 
+ * @param t_redirect_single_commands_params *params : structure for
+ 	single_command parameters
+ 	***env: *** is called in the calling function
+ 	needed ** to free data if child process exits.
+ * 
+ * @return: -1 if forking fails. 0 if the whole execution is successful
+ */
+
+int	handle_fork_plus_executing_child(t_redirect_single_command_params
+*params, char ***env)
+{
 	params->pid = fork();
 	if (params->pid < 0)
 	{
@@ -74,79 +148,43 @@ int	handle_redirects(t_redirect_single_command_params *params,
 		return (-1);
 	}
 	if (params->pid == 0)
-		execute_child_process_for_redirections(params, env, input);
+		child_process_other_cases(params, env);
 	else
 		wait(NULL);
 	return (0);
 }
 
 /**
- * @function: execute_child_process_for_redirections
- * @brief: further summarising into a cleaner structure to read
+ * @function: handle_single_commands
+ * @brief: handles single commands with built-in features
  * 
- * @param t_redirect_single_command_params *params : structure to
- 	store parameters for handling redirects / no redirects
- 	char ***env : *** is called in calling function
- 		only needed ** for freeing in child process
- 	char *input : initial readline from main function
- 		used for freeing purposes if program fails
+ * @param t_redirect_single_command_params *params : structure for
+ 	single_command parameters
+ 	***env : needed *** to amend env
+ 	if cd and export command is called
+ 	needed ** to free data if child process exits.
  * 
- * @return: void function
+ * @return: -1 if forking process fails, 0 if all commands are successfully run
  */
 
-void	execute_child_process_for_redirections(t_redirect_single_command_params
-*params, char ***env, char *input)
-
-{
-	handle_redirections_file_opening(params, env, input);
-	handle_dup_and_closing_fd(params, env, input);
-	handle_execve_for_redirections(params, env, input);
-}
-
-/**
- * @function: handle_single_commands_no_heredocs
- * @brief: handling forking process if no heredocs is found
- * 
- * @param t_redirect_single_command_params *params : structure to
- 	store parameters for handling redirects / no redirects
- 	char ***env : *** is called in calling function
- 		only needed ** for freeing in child process
- 	char *input : initial readline from main function
- 		used for freeing purposes if program fails
- * 
- * @return: -1 if failure, 0 if success
- */
-
-int	handle_single_commands_no_heredocs(t_redirect_single_command_params *params,
-	char ***env, char *input)
+int	handle_single_commands(t_redirect_single_command_params *params,
+	char ***env)
 {	
-	if (handle_fork_plus_executing_child(params, env, input) == -1)
-		return (-1);
-	return (0);
-}
-
-/**
- * @function: handle_other_cases
- * @brief: handling other test case scenarios
- * 
- * @param t_redirect_single_command_params *params : structure to
- 	store parameters for handling redirects / no redirects
- 	char ***env : *** is needed to modify env variable if needed
- 	char *input : initial readline from main function
- 		used for freeing purposes if program fails
- * 
- * @return: -1 if forking fails 0 if all commands is executed
- */
-
-int	handle_other_cases(t_redirect_single_command_params *params,
-	char ***env, char *input)
-
-{
-	params->av = params->result->cmd;
-	params->ac = 0;
-	while (params->av[params->ac] != NULL)
-		params->ac++;
-	if (handle_single_commands(params, env, input) == -1)
+	if (ft_strcmp(params->av[0], "echo") == 0)
+		echo_command(params->ac, params->av);
+	else if (ft_strcmp(params->av[0], "cd") == 0)
+		cd_command(params->ac, params->av, env);
+	else if (ft_strcmp(params->av[0], "pwd") == 0)
+		pwd_command(params->ac, params->av);
+	else if (ft_strcmp(params->av[0], "export") == 0)
+		export_command(params->ac, params->av, env);
+	else if (ft_strcmp(params->av[0], "unset") == 0)
+		unset_command(params->ac, params->av, *env);
+	else if (ft_strcmp(params->av[0], "env") == 0)
+		env_command(params->ac, params->av, *env);
+	else if (ft_strcmp(params->av[0], "exit") == 0)
+		exit_command(params, *env);
+	else if (handle_fork_plus_executing_child(params, env) == -1)
 		return (-1);
 	return (0);
 }
