+++ 06_Built-Ins/exit.c
@@ -19,16 +19,14 @@
  * @param **av: to free up the arguments which
  		are already split into tokens
  	 **env: to free up the duplicated env
- 	 *input: to free up readline input variable
  * 
  * @return: void function
  */
 
 static void	exit_is_the_only_argument(t_redirect_single_command_params *params,
-		char **env, char *input)
+		char **env)
 
 {
-	free(input);
 	free_dup_envp(env);
 	rl_clear_history();
 	ft_printf("exit\n");
@@ -44,21 +42,19 @@ static void	exit_is_the_only_argument(t_redirect_single_command_params *params,
  * @param **av: to free up the arguments which
  		are already split into tokens
  	 **env: to free up the duplicated env
- 	 *input: to free up readline input variable
  * 
  * @return: void function
  */
 
 static void	exit_with_one_other_argument(
 		t_redirect_single_command_params *params,
-		char **env, char *input)
+		char **env)
 
 {
 	int	exit_status;
 
 	exit_status = 0;
 	exit_status = ft_atoi(params->av[1]);
-	free(input);
 	free_dup_envp(env);
 	rl_clear_history();
 	ft_printf("exit\n");
@@ -104,19 +100,17 @@ static int	is_argument_numeric(const char *arg)
  	**av: the strings belonging to each argument.
  	**env: duplicated env. being called so
  		i can free it and exit from here.
- 	*input: input from readline. being called 
- 		so i can free it and exit from here.
  * 
  * @return: void function
  */
 
 void	exit_command(t_redirect_single_command_params *params,
-		char **env, char *input)
+		char **env)
 {
 	if (params->ac > 0 && ft_strncmp(params->av[0], "exit", 4) == 0)
 	{
 		if (params->ac == 1 && ft_strlen(params->av[0]) == 4)
-			exit_is_the_only_argument(params, env, input);
+			exit_is_the_only_argument(params, env);
 		else if (params->ac == 1 && ft_strlen(params->av[0]) != 4)
 			ft_dprintf(2, "%s: command not found\n", params->av[0]);
 		else if (params->ac == 2 && ft_strlen(params->av[0]) == 4)
@@ -125,14 +119,14 @@ void	exit_command(t_redirect_single_command_params *params,
 			{
 				ft_dprintf(2, "%s: %s: numberic"
 					" argument required\n", params->av[0], params->av[1]);
-				exit_is_the_only_argument(params, env, input);
+				exit_is_the_only_argument(params, env);
 			}
-			exit_with_one_other_argument(params, env, input);
+			exit_with_one_other_argument(params, env);
 		}
 		else
 		{
 			ft_dprintf(2, "%s: too many arguments\n", params->av[0]);
-			exit_is_the_only_argument(params, env, input);
+			exit_is_the_only_argument(params, env);
 		}
 	}
 }
