+++ Makefile
@@ -10,7 +10,7 @@ LEXER_FILES = lexer.c	tokenize.c	lexer_utils.c	ft_split_ignore_quotes.c	ft_strch
 EXPANSION_FILES = expansion.c	expansion_utils_a.c	expansion_utils_b.c	expansion_utils_c.c
 PARSER_FILES = parser.c	parser_utils_a.c	parser_utils_b.c	parser_rd_helper.c	parser_rd_helper_b.c	parser_recursive_descent.c	parser_print_tree.c	parser_ast_to_ll.c	parser_exec.c
 REDIRECTION_FILES = path_construction.c	path_construction_utils.c	piping.c	redirections_utils.c	redirections_utils1.c	redirections_utils2.c	redirections_utils4.c
-EXECUTION_FILES = single_command.c	single_command_utils1.c	single_command_utils2.c	single_command_utils3.c	single_command_utils4.c	multiple_commands.c	multiple_commands_utils1.c	multiple_commands_utils2.c	multiple_commands_utils3.c	multiple_commands_utils4.c	multiple_commands_utils5.c	multiple_commands_utils6.c	multiple_commands_utils7.c	heredocs.c	heredocs_utils1.c	heredocs_utils2.c	free_pipes.c
+EXECUTION_FILES = single_command.c	single_command_utils1.c	single_command_utils2.c	single_command_utils3.c	single_command_utils4.c	single_command_utils5.c	multiple_commands.c	multiple_commands_utils1.c	multiple_commands_utils2.c	multiple_commands_utils3.c	multiple_commands_utils4.c	multiple_commands_utils5.c	multiple_commands_utils6.c	multiple_commands_utils7.c	heredocs.c	heredocs_utils1.c	heredocs_utils2.c	free_pipes.c
 BUILT_INS_FILES = cd.c	updating_env_after_cd.c	copy_envp.c	copy_envp_with_ac.c env.c	free_dup_envp.c	exit.c	pwd.c	free_tokens.c	echo.c	unset.c	export.c	export_utils1.c	export_utils2.c	export_utils3.c	export_utils4.c	export_utils5.c	export_utils6.c	export_utils7.c	ft_strcmp.c
 SIGNAL_FILES = signal_handler.c	signal.c
