+++ includes/built_in.h
@@ -43,8 +43,7 @@ char	**cd_command(int ac, char **av, char ***env);
 char	**updating_env(char ***env, char *old_pwd);
 void	pwd_command(int ac, char **av);
 void	env_command(int ac, char **av, char **env);
-void	exit_command(t_redirect_single_command_params *params,
-			char **env, char *input);
+void	exit_command(t_redirect_single_command_params *params, char **env);
 void	echo_command(int ac, char **av);
 char	**unset_command(int ac, char **av, char **env);
 char	**export_command(int ac, char **av, char ***env);
