+++ minishell.c
@@ -95,10 +95,10 @@ int	main(int ac, char **av, char **envp)
 			ft_free(&token_data, &ast_root);
 			continue ;
 		}
-		ft_free(&token_data, &ast_root);
-		execution(exec_data, &env, input);
-		execution_with_pipes(exec_data, &env, input);
 		free(input);
+		ft_free(&token_data, &ast_root);
+		execution(exec_data, &env);
+		execution_with_pipes(exec_data, &env);
 		ft_lstclear(&exec_data, ft_free_exec_data);
 	}
 	free_dup_envp(env);
